Move `DependencyGraph` from `//language-support/scala/codegen` to `//language-support/codegen-common/`
changelog_begin
changelog_end

The idea is that of re-using `DependencyGraph` for the Java codegen.

Builds on top of https://github.com/digital-asset/daml/pull/13361 and
contributes towards https://github.com/digital-asset/daml/issues/13324.
