+++ language-support/codegen-common/BUILD.bazel
@@ -14,12 +14,14 @@ da_scala_library(
         "@maven//:com_github_scopt_scopt",
         "@maven//:io_circe_circe_core",
         "@maven//:org_typelevel_cats_core",
+        "@maven//:org_scalaz_scalaz_core",
     ],
     tags = ["maven_coordinates=com.daml:codegen-jvm-common:__VERSION__"],
     visibility = ["//visibility:public"],
     deps = [
         "//daml-assistant/scala-daml-project-config",
         "//daml-lf/data",
+        "//daml-lf/interface",
         "//libs-scala/build-info",
         "@maven//:ch_qos_logback_logback_classic",
     ],
