+++ language-support/codegen-common/src/main/scala/com/digitalasset/daml/codegen/dependencygraph/DependencyGraph.scala
@@ -3,12 +3,11 @@
 
 package com.daml.codegen.dependencygraph
 
-import com.daml.lf.iface._
-import com.daml.lf.data.Ref.Identifier
-import com.daml.codegen.lf.LFUtil
-import com.daml.codegen.lf
+import com.daml.codegen.Util
 import com.daml.lf.data.ImmArray.ImmArraySeq
-import lf.DefTemplateWithRecord
+import com.daml.codegen.lf.DefTemplateWithRecord
+import com.daml.lf.data.Ref.Identifier
+import com.daml.lf.iface.{EnvironmentInterface, InterfaceType}
 import scalaz.std.list._
 import scalaz.syntax.bifoldable._
 import scalaz.syntax.foldable._
@@ -25,15 +24,15 @@ private[codegen] object DependencyGraph {
           qualName,
           Node(
             TypeDeclWrapper(typeDecl),
-            typeDecl.bifoldMap(LFUtil.genTypeTopLevelDeclNames)(LFUtil.genTypeTopLevelDeclNames),
+            typeDecl.bifoldMap(Util.genTypeTopLevelDeclNames)(Util.genTypeTopLevelDeclNames),
             collectDepError = false,
           ),
         )
       }
     val templateNodes =
       decls.to(ImmArraySeq).collect { case (qualName, InterfaceType.Template(typ, tpl)) =>
-        val recDeps = typ.foldMap(LFUtil.genTypeTopLevelDeclNames)
-        val choiceAndKeyDeps = tpl.foldMap(LFUtil.genTypeTopLevelDeclNames)
+        val recDeps = typ.foldMap(Util.genTypeTopLevelDeclNames)
+        val choiceAndKeyDeps = tpl.foldMap(Util.genTypeTopLevelDeclNames)
         (
           qualName,
           Node(
