+++ language-support/codegen-common/src/main/scala/com/digitalasset/daml/codegen/exception/UnsupportedTypeError.scala
