+++ language-support/scala/codegen/BUILD.bazel
@@ -47,6 +47,7 @@ da_scala_library(
         "//daml-lf/archive:daml_lf_archive_reader",
         "//daml-lf/data",
         "//daml-lf/interface",
+        "//language-support/codegen-common",
         "@maven//:org_slf4j_slf4j_api",
     ],
 )
